Remove the spurious CR characters
This situation happens when running the test on Windows.

When wait-for is copied to the docker image, it has appended the
carriage return character by the end of each line.
This rendered the wait-for unusable by showing a Bad interpreter message
(aka status code 126)

It can be resolved by replacing carriage return character with line breaks
instead.

Reference: https://askubuntu.com/a/305001